Improved scheduleAlerts
Alerts are now being scheduled when added / updated and cancelled when deleted rather than cancelling and scheduling all alerts when entering ListFragment
